requirements.txt
google-search-results>=2.4.2
apscheduler>=3.10.4
python-dotenv>=1.0.0
SQLAlchemy>=2.0
flask>=3.0.0
gunicorn>=21.0.0
pytest>=7.0.0
